string_helpers: Backport from v5.13-rc6
Summary:

This is a self-contained backport of string_helpers and its tests from
v5.13-rc6, which the printk patches are based on. This is required to support
ESCAPE_APPEND (which we use later), and set the base for the new quoting
behaviour in commit "string_helpers: Escape double quotes in escape_special".

The only additional change is to revert
815c913a0dc94b75c4b5bcb850817d912f5d9400, which removes
*_escape_mem_ascii, which is needed for 5.12 in general.

Test Plan:

Tested as part of larger merge.

Reviewers: jweiner, alastorze, kernel_ops

Tasks: T99779272