Revert "block: kill readahead after throttling"
This reverts commit d2d464ec885c5dd6968255782fadb0805f663fa2.

This was forward ported from v5.6, but introduced an accounting bug
because the surrounding code has changed.  We saw deadlocks with
everyone piling up waiting for a frozen nvme queue:

PID: 503387     power_estimatio D

The freezer was waiting for all the IO to finish:

Omar figured out the IO counters just went wrong because this commit
expected the caller to drop the counter and it no longer does so.  It's
a oneline fix, but it turns out this commit from upstream does
everything we need:

ca47e8c72ae141587cabf3dab693f6754d8c416b
Author: Josef Bacik <jbacik@fb.com>
Date:   Tue Jul 3 11:15:03 2018 -0400

    mm: skip readahead if the cgroup is congested

So, reverting instead.

Signed-off-by: Chris Mason <clm@fb.com>