bpf: don't break identity mapping hugepages when loading programs
When bpf programs are loaded, permissions of the text sections must be
changed to be executable and read-only. But the vmalloc-mapped text
sections have aliases through the identity mapping, which are changed
the same way to ensure there are no writable aliases to exec memory.

But over time, loading and unloading bpf programs fragments the
kernel's identity mapping, which starts out with gigabyte pages, into
4k pages. This gradually increases TLB pressure on in-kernel
execution, causing a measurable regression in application performance.

Switch bpf to new noalias versions of the permission change helpers,
so that the identity mapping is left alone entirely.

BPF is using the vmalloc addresses to access and jump into this
memory, and nobdoy *should* be using the identity mapping aliases in
any way. But it does mean there is a writeable mapping to executable
bpf code.